requirements.txt
langchain_community==0.3.31
langchain_core==0.3.78
langchain_groq==0.3.8
langchain_huggingface==0.3.1
python-dotenv==1.1.1
Requests==2.32.5
streamlit==1.50.0
pypdf==6.1.1
sentence-transformers==5.1.1
chromadb==1.1.1